Make CloudStorageTools::createUploadUrl compatible with PHP7
`createUploadUrl` calls `getUploadMaxFileSizeInBytes`, which does the following operation on a mostly-numeric string:

`$val *= 1024`

In PHP5, this works fine, because non-numerica value is stripped from the end.

In PHP7, a warning is issued which states `A non well formed numeric value encountered`.

The value has a final non-numeric character which can be stripped to prevent this warning.

This will continue to function in PHP5 as well.

Fixes #10